restructure program for pip distribution
